feat(ccwc): implementing default commands for ccwc
-Adding c, l, w, m commands to use with bufio scanner
